Refactored message/formatter creation.
Extracted SlackFormatter into its own file, and created it outside of
the Message constructor, to avoid the (incorrect) appearance of creating
a million independent SlackFormatter objects when actually they all just
wrap the same dicts.
